create state_yaml; update check_yaml.
Fixed check_yaml so that if we put a state that's a legitimate state that we don't have data for (e.g., texas as of writing), it'll kick back that we don't have it in our records. Made check_yaml so that people could check the yaml data that we currently have for their state of interest.